Fast build

+++ app/Http/Controllers/PostsController.php
@@ -4,9 +4,20 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Post;
+use Illuminate\Support\Facades\storage;
 
 class PostsController extends Controller
 {
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'about', 'show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -14,7 +25,7 @@ class PostsController extends Controller
      */
     public function index()
     {
-        $posts = Post::Orderby('created_at', 'desc')->paginate(10);
+        $posts = Post::Orderby('created_at', 'desc')->paginate(9);
         return view('posts.index')->with('posts', $posts);
     }
 
@@ -39,13 +50,32 @@ class PostsController extends Controller
         //Create Post
         $this->validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        //Cover image function
+
+        if($request->hasFile('cover_image')){
+            //get file name
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //path file
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            //get extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            //filename store
+            $fileNameToStore= $filename.'_'.time().'.'.$extension;
+            //upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        } else {
+            $fileNameToStore = 'noimage.jpg';
+        }
         
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->body;
         $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Created');
@@ -72,6 +102,10 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        if(auth()->user()->id !==$post->user_id){
+            return redirect('/posts')->with('error','Unauthorized Page');
+        }
         return view('posts.edit')->with('post',$post);
     }
 
@@ -89,10 +123,28 @@ class PostsController extends Controller
             'title' => 'required',
             'body' => 'required'
         ]);
+
+        //Cover image function
+
+        if($request->hasFile('cover_image')){
+            //get file name
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            //path file
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            //get extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            //filename store
+            $fileNameToStore= $filename.'_'.time().'.'.$extension;
+            //upload image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        }
         
         $post = Post::find($id);
         $post->title = $request->input('title');
         $post->body = $request->body;
+        if($request->hasFile('cover_image')){
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Updated');
@@ -107,6 +159,14 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        if(auth()->user->id !==$post->user_id()){
+            return redirect('/posts')->with('error','Unauthorized Page');
+        }
+
+        if($post->cover_image != 'noimage.jpg'){
+            // Delete Image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
         $post->delete();
 
         return redirect('/posts')->with('success', 'Post Deleted');
